requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
contourpy==1.1.0
croniter==1.3.15
cycler==0.11.0
Cython==0.29.35
dateutils==0.6.12
deepdiff==6.3.0
efficientnet-pytorch==0.7.1 
exceptiongroup==1.1.1
fastapi==0.98.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
h11==0.14.0
huggingface-hub==0.15.1
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lightning==2.0.4
lightning-cloud==0.5.37
lightning-sam @ file:///home/someone/stage_jonathan/lightning-sam/lightning-sam
lightning-utilities==0.9.0
lit==16.0.6
lvis==0.5.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
munch==3.0.0
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
ordered-set==4.1.0
packaging==23.1
Pillow==9.3.0
pretrainedmodels==0.7.4
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
python-box==7.0.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.4
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
segment-anything==1.0
segmentation-models-pytorch==0.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
timm==0.9.2
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.0
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
